+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Changelog
 
+- New transformation functions are added: `radiansToDegrees`, `degreesToRadians`, `radiansLat`, `radiansLong`.
+
 ### 2.0.3 (2018-09-15)
 
 - `satrec` object is not cloned in `sgp4` and `sgp4init` functions due to [performance reasons](https://github.com/shashwatak/satellite-js/issues/45).